backend/pyproject.toml
[tool.poetry]
name = "DOBOKU_Post"
version = "0.1.0"
description = ""
authors = ["saka-21 <kasaikasai572@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.3"
Django = "3.0"
django-cleanup = "^5.0.0"
djangorestframework = "3.10"
django-imagekit = "^4.0.2"
pillow = "^7.2.0"
django-cors-headers = "^3.4.0"
djoser = "^2.0.3"
django-filter = "^2.3.0"
psycopg2 = "^2.8.5"
djangorestframework-simplejwt = "^4.4.0"
django-environ = "^0.4.5"
uwsgi = "^2.0.19"
boto3 = "^1.14.56"
django-storages = "^1.10"
environ = "^1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.5.3"
autopep8 = "^1.5.4"
flake8 = "^3.8.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
